backend/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yorumator-backend"
version = "0.1.0"
description = "FastAPI backend for the Yorumator decision support platform"
requires-python = ">=3.11"
readme = "README.md"
authors = [{name = "Yorumator Team"}]
dependencies = [
  "fastapi>=0.115.0,<0.116.0",
  "uvicorn[standard]>=0.30.0,<0.31.0",
  "sqlalchemy>=2.0.31,<2.1.0",
  "psycopg[binary]>=3.2.1,<3.3.0",
  "passlib[argon2]>=1.7.4,<1.8.0",
  "python-jose[cryptography]>=3.3.0,<3.4.0",
  "pyotp>=2.9.0,<3.0.0",
  "python-multipart>=0.0.9,<0.1.0",
  "pydantic-settings>=2.4.0,<2.5.0",
  "redis>=5.0.8,<5.1.0",
  "httpx>=0.27.0,<0.28.0",
  "celery>=5.4.0,<5.5.0",
  "boto3>=1.34.138,<1.35.0",
  "alembic>=1.13.1,<1.14.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "ruff",
  "mypy",
  "types-redis",
]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP"]
